+++ src/mappings/Season2020/Drivers/NumberByName.js
@@ -1,0 +1,8 @@
+const ByNumber = require('./ByNumber');
+
+const NumberByName = {};
+for (const [number, driver] of Object.entries(ByNumber)) {
+    NumberByName[driver.name] = number;
+}
+
+module.exports = NumberByName;