+++ src/mappings/Season2020/Models/IdByName.js
@@ -1,0 +1,4 @@
+const reverseMapping = require('../../utils/reverseMapping');
+const NameById = require('./NameById');
+const IdByName = reverseMapping(NameById);
+module.exports = IdByName;