+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.1.0 (30/04/2020)
+
+### New Features
+* Added `collections` with support for inventory metadata layouts.
+* Added season 2020 mappings (WIP).
+
+### Chores
+ * Moved `metadataFromId.js` to `scripts` folder.
+
 ## 0.0.2 (23/04/2020)
 
 ### Bugfixes